+++ R/find_data.R
@@ -1,6 +1,7 @@
 #' @title Extract data from a model object
 #' @description Find a model frame in a model object or try to reconstruct one
 #' @param model The model object.
+#' @param \dots Additional arguments passed to methods.
 #' @param env An environment in which to look for the \code{data} argument to the modelling call.
 #' @return A data.frame, typically with one column unless the variable is a factor with more than two levels.
 #' @examples
@@ -10,20 +11,19 @@
 #' 
 #' @seealso \code{\link{prediction}}
 #' @export
-find_data <- function(model, env = parent.frame()) {
-  UseMethod("find_data")
-  
+find_data <- function(model, ...) {
+    UseMethod("find_data")
 }
 
 #' @rdname find_data
 #' @export
-find_data.default <- function(model, env = parent.frame()) {
-  if (!is.null(model[["call"]][["data"]])) {
-    data <- eval(model[["call"]][["data"]], env) 
-  } else { 
-    data <- get_all_vars(model[["terms"]], data = model[["model"]])
-  }
-  data
+find_data.default <- function(model, env = parent.frame(), ...) {
+    if (!is.null(model[["call"]][["data"]])) {
+        data <- eval(model[["call"]][["data"]], env) 
+    } else { 
+        data <- get_all_vars(model[["terms"]], data = model[["model"]])
+    }
+    data
 }
 
 #' @rdname find_data
@@ -36,9 +36,13 @@ find_data.glm <- find_data.default
 
 #' @rdname find_data
 #' @export
-find_data.svyglm <- function(model, env = parent.frame()) {
+find_data.svyglm <- function(model, ...) {
+    data <- model[["data"]]
+    data
+}
 
-  data <- model$data
-
-  data
+#' @rdname find_data
+#' @export
+find_data.merMod <- function(model, env = parent.frame(), ...) {
+    model.frame(model)
 }